+++ android/src/main/java/com/sfourdrinier/unifiedblemanager/adapter/BleAdapterCreator.java
@@ -1,7 +1,0 @@
-package com.sfourdrinier.unifiedblemanager.adapter;
-
-import android.content.Context;
-
-public interface BleAdapterCreator {
-  BleAdapter createAdapter(Context context);
-}