+++ android/src/main/java/com/sfourdrinier/unifiedblemanager/adapter/OnErrorCallback.java
@@ -1,8 +1,0 @@
-package com.sfourdrinier.unifiedblemanager.adapter;
-
-import com.sfourdrinier.unifiedblemanager.adapter.errors.BleError;
-
-public interface OnErrorCallback {
-
-  void onError(BleError error);
-}