+++ android/src/main/java/com/sfourdrinier/unifiedblemanager/adapter/OnEventCallback.java
@@ -1,6 +1,0 @@
-package com.sfourdrinier.unifiedblemanager.adapter;
-
-public interface OnEventCallback<T> {
-
-  void onEvent(T data);
-}