+++ android/src/main/java/com/sfourdrinier/unifiedblemanager/adapter/OnSuccessCallback.java
@@ -1,6 +1,0 @@
-package com.sfourdrinier.unifiedblemanager.adapter;
-
-public interface OnSuccessCallback<T> {
-
-  void onSuccess(T data);
-}