+++ ios/BlePlxDebugLogging.h
@@ -1,6 +1,0 @@
-#import <Foundation/Foundation.h>
-
-FOUNDATION_EXPORT BOOL BlePlxDebugLoggingEnabled(void);
-
-FOUNDATION_EXPORT void BlePlxDebugLog(NSString *format, ...) NS_FORMAT_FUNCTION(1, 2);
-